docs(front/basic): document token helpers
